Refactored Validation
