+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/EventValidationService.java
@@ -6,7 +6,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseData;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseDetails;
-import uk.gov.hmcts.ecm.common.model.ccd.items.DepositTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.JudgementTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.JurCodesTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.RepresentedTypeRItem;
@@ -35,7 +34,6 @@ import static java.util.stream.Collectors.joining;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLAIMANT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLOSING_HEARD_CASE_WITH_NO_JUDGE_ERROR;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLOSING_LISTED_CASE_ERROR;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.DEPOSIT_REFUNDED_GREATER_DEPOSIT_ERROR;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.DUPLICATED_JURISDICTION_CODES_JUDGEMENT_ERROR;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.DUPLICATE_JURISDICTION_CODE_ERROR_MESSAGE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.EARLY_DATE_RETURNED_FROM_JUDGE_ERROR_MESSAGE;
@@ -60,7 +58,6 @@ import static uk.gov.hmcts.ecm.common.model.helper.Constants.RESPONDENT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.RESP_REP_NAME_MISMATCH_ERROR_MESSAGE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.SUBMITTED_STATE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.TARGET_HEARING_DATE_INCREMENT;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.UNABLE_TO_FIND_PARTY;
 import static uk.gov.hmcts.ethos.replacement.docmosis.helpers.Helper.getActiveRespondents;
 
 @Slf4j
@@ -349,71 +346,6 @@ public class EventValidationService {
                     errors.add(MISSING_JUDGEMENT_JURISDICTION_MESSAGE);
                 }
                 break;
-            }
-        }
-    }
-
-    public List<String> validateDepositOrder(CaseData caseData) {
-        List<String> errors = new ArrayList<>();
-        if (!CollectionUtils.isEmpty(caseData.getDepositCollection())) {
-
-            for (DepositTypeItem depositTypeItem : caseData.getDepositCollection()) {
-                if (!isNullOrEmpty(depositTypeItem.getValue().getDepositAmountRefunded())
-                        && (isNullOrEmpty(depositTypeItem.getValue().getDepositAmount())
-                        || Integer.parseInt(depositTypeItem.getValue().getDepositAmountRefunded())
-                        > Integer.parseInt(depositTypeItem.getValue().getDepositAmount()))) {
-                    errors.add(DEPOSIT_REFUNDED_GREATER_DEPOSIT_ERROR);
-                }
-
-                validateDepositOrderAgainst(errors, depositTypeItem);
-                validateDepositRequestedBy(errors, depositTypeItem);
-                validateDepositRefunded(errors, depositTypeItem);
-            }
-        }
-        return errors;
-    }
-
-    private void validateDepositRefunded(List<String> errors, DepositTypeItem depositTypeItem) {
-        if (!isNullOrEmpty(depositTypeItem.getValue().getDepositRefund())
-                && depositTypeItem.getValue().getDynamicDepositRefundedTo() != null) {
-            var dynamicRefundedTo = depositTypeItem.getValue().getDynamicDepositRefundedTo()
-                    .getValue().getCode();
-            if (dynamicRefundedTo.startsWith("R: ")) {
-                depositTypeItem.getValue().setDepositRefundedTo(RESPONDENT_TITLE);
-            } else if (dynamicRefundedTo.startsWith("C: ")) {
-                depositTypeItem.getValue().setDepositRefundedTo(CLAIMANT_TITLE);
-            } else {
-                errors.add(UNABLE_TO_FIND_PARTY);
-            }
-        }
-    }
-
-    private void validateDepositRequestedBy(List<String> errors, DepositTypeItem depositTypeItem) {
-        if (depositTypeItem.getValue().getDynamicDepositRequestedBy() != null) {
-            var dynamicRequestedBy = depositTypeItem.getValue().getDynamicDepositRequestedBy()
-                    .getValue().getCode();
-            if (dynamicRequestedBy.startsWith("R:")) {
-                depositTypeItem.getValue().setDepositRequestedBy(RESPONDENT_TITLE);
-            } else if (dynamicRequestedBy.startsWith("C:")) {
-                depositTypeItem.getValue().setDepositRequestedBy(CLAIMANT_TITLE);
-            } else if (dynamicRequestedBy.equals("Tribunal")) {
-                depositTypeItem.getValue().setDepositRequestedBy("Tribunal");
-            } else {
-                errors.add(UNABLE_TO_FIND_PARTY);
-            }
-        }
-    }
-
-    private void validateDepositOrderAgainst(List<String> errors, DepositTypeItem depositTypeItem) {
-        if (depositTypeItem.getValue().getDynamicDepositOrderAgainst() != null) {
-            var dynamicOrderAgainst = depositTypeItem.getValue().getDynamicDepositOrderAgainst()
-                    .getValue().getCode();
-            if (dynamicOrderAgainst.startsWith("R:")) {
-                depositTypeItem.getValue().setDepositOrderAgainst(RESPONDENT_TITLE);
-            } else if (dynamicOrderAgainst.startsWith("C:")) {
-                depositTypeItem.getValue().setDepositOrderAgainst(CLAIMANT_TITLE);
-            } else {
-                errors.add(UNABLE_TO_FIND_PARTY);
             }
         }
     }
